fix database.

+++ WeatherForecastApp/Models/OpenWheatherMapResponse/Coord.cs
@@ -3,6 +3,6 @@
     public class Coord
     {
         public double lon { get; set; }
-        public int lat { get; set; }
+        public double lat { get; set; }
     }
 }
